tidying up code, hover states for carousel and error messages
